# Conflicts:
#	src/main/java/com/example/demo/controller/QuizController.java

+++ src/main/java/com/example/demo/vo/ResponseQuizDetail.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 @Builder
 public class ResponseQuizDetail {
     private Long quizNum;
-    private Integer answerRate;
+    private Float answerRate;
     private Integer trialUserCount;
     private Integer answerUserCount;
